requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
sqlparse==0.5.5

pillow~=12.1.0
pip~=24.0
attrs~=25.4.0
redis~=7.1.0
amqp~=5.3.1
vine~=5.1.0
PyYAML~=6.0.3
typing_extensions~=4.15.0
kombu~=5.6.2
billiard~=4.2.4
packaging~=26.0
celery~=5.6.2
click~=8.3.1
python-dateutil~=2.9.0.post0
tzlocal~=5.3.1
wcwidth~=0.3.2
six~=1.17.0
prompt_toolkit~=3.0.52
jsonschema~=4.26.0
referencing~=0.37.0
uritemplate~=4.2.0
inflection~=0.5.1
PyJWT~=2.10.1
python-dotenv~=1.2.1